refactor: move write options to write/options.rs
