+++ src/read/zipfile.rs
@@ -4,6 +4,7 @@ use crate::CompressionMethod;
 use crate::DateTime;
 use crate::HasZipMetadata;
 use crate::ZIP64_BYTES_THR;
+use crate::format::ffi;
 use crate::read::ExtraField;
 use crate::read::RootDirFilter;
 use crate::read::make_writable_dir_all;
@@ -11,7 +12,8 @@ use crate::read::readers::{ZipFileReader, ZipFileSeekReader};
 use crate::result::ZipResult;
 use crate::result::invalid;
 use crate::types::ZipFileData;
-use crate::types::{SimpleFileOptions, ffi};
+use crate::write::options::SimpleFileOptions;
+
 use core::mem::replace;
 use std::borrow::Cow;
 use std::ffi::OsStr;
@@ -320,7 +322,7 @@ impl<'a, R: Read + ?Sized> ZipFile<'a, R> {
         options.normalize();
         #[cfg(feature = "aes-crypto")]
         if let Some((mode, vendor_version)) = self.get_metadata().aes_settings() {
-            use crate::types::EncryptWith;
+            use crate::write::options::EncryptWith;
             // Preserve AES metadata in options for downstream writers.
             // This is metadata-only and does not trigger encryption.
             options.encrypt_with = Some(EncryptWith::Aes {
